+++ Interfaces/Permissionner.sol
@@ -1,4 +1,4 @@
 contract Permissionner {
-    function perms(address addr) constant returns (uint8) { }
+    function perms(address addr) constant returns (uint) { }
     function setPermission(address addr, uint perm) returns (bool);
 }
